Sound Fix, High Score Background Change, WBC
Fixed the issue with the repetitive clicking sound when moving selected
viruses.
Changed the background of the High Score Menu to match the rest of the
game
Fixed an issue with WBC going off camera by locking them to a single
plane of movement.
There is an issue currently with the hitbox of WBC.