Fixed WhiteBloodCell Hitbox Issue
The collider was getting off because of the ridiculous scaling on the
WBC. Apparently Unity has a bug in which it defaults the FBX scaling to
0.01. Bumped up the scaling on the model to reduce the effective scaling
on the prefab. This seemingly fixed the hitbox problem. Also reduced the
intensity of the CellBurstPrefab particle effect. I was getting
complaints from testers that it was too distracting and obscuring.